+++ test/mssql/listing1-22.sql
@@ -1,6 +1,0 @@
-
-SELECT orderid, customerid,
-  COUNT(*) OVER(PARTITION BY customerid) AS num_orders
-FROM dbo.Orders
-WHERE customerid IS NOT NULL
-  AND orderid % 2 = 1;