+++ test/mssql/listing1-23.sql
@@ -1,6 +1,0 @@
-
-SELECT orderid, customerid
-FROM dbo.Orders
-WHERE customerid IS NOT NULL
-  AND orderid % 2 = 1
-ORDER BY COUNT(*) OVER(PARTITION BY customerid) DESC;