+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added params to SQLite attached database: alasql('ATTACH SQLITE DATABASE a(?)',[event],cb);
 * Root directory was cleaned
 * Gulp version is updated
+* Fixed bug with (SELECT) and EXISTS() in SELECTS with GROUP BY
 
 ### 0.0.43 "The Wall" (25.03.2015 - 01.04.2015)
 * Created "develop" branch for git-flow